pyproject.toml
[project]
name = "svcalc-modified"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "urasaku77", email = "okadaisyskou@yahoo.co.jp" }
]
dependencies = [
    "jaconv",
    "opencv-python",
    "pandas",
    "pillow",
    "pyocr",
    "simpleobsws",
    "ttkthemes",
]
readme = "README.md"
requires-python = ">= 3.10"

[tool.rye.scripts]
main = "python main.py"
home = "python home/home.py"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "mypy>=1.8.0",
    "ruff>=0.1.13",
    "pre-commit>=3.6.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/svcalc_modified"]

[tool.ruff]
target-version = "py310"

select = ["E", "F", "W", "I", "B"]
fixable = ["ALL"]

ignore = ["E402", "E501", "E722"]
